+++ src/config/config-data.ts
@@ -13,10 +13,11 @@ export interface ConfigData {
 	 * - "html5" (default) for the builtin metadata.
 	 * - node module which export metadata
 	 * - local path to json or js file exporting metadata.
+	 * - object with inline metadata
 	 *
 	 * If elements isn't specified it defaults to `["html5"]`
 	 */
-	elements?: string[];
+	elements?: (string|object)[];
 
 	/**
 	 * List of plugins.