+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Upcoming release
 
+### Features
+
+- Support writing element metadata inline in configuration file.
+
 ## 0.12.0 (2018-10-17)
 
 ### Features